+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Single class bug fixed
+- `CLASS_NUMBER_ERROR` error type changed to `pycmMatrixError`
 ## [2.7] - 2020-05-11
 ### Added
 - `average` method